Added Solo.py Script and Other Changes
Added Solo.py Script to control the Drone and also changed the connection to the DB on the Listener Script. I believe they had it wrong and I just fixed it but I can't tell until we test it. Also I do not think that they were connected to the DB at all.